refactor: :recycle: собрать рекурсивные преобразователи MetadataValue
Составные значения теперь рекурсивно обрабатываются единым ядром каждого направления. Старые пути fixedArray и formChoiceList сохранены как реэкспорты.
