+++ packages/core/metadata/commonObjects/metadataValue/toYAML.ts
@@ -1,12 +1,13 @@
 import { ConfigurationContext } from "../../context/types"
 import type { PropertyRule } from "../../orchestration/property/types"
 import { registerTypeRule } from "../../orchestration/property/typeRuleRegistry"
-import { exportFixedArrayToYAML } from "./fixedArray/toYAML"
-import { exportFormChoiceListToYAML } from "./formChoiceList/toYAML"
 import { primitiveValueHandlers } from "./handlers"
 import { exportStandardPeriodToYAML } from "../standardPeriod/toYAML"
+import { DataCompositionComparisonTypeToYAML } from "../../systemEnumerations/types"
+import { exportI8nTextToYAML } from "../i8nText/toYAML"
 import {
   MetadataFixedArrayValue,
+  MetadataFixedArrayValueYAML,
   MetadataFormChoiceListValue,
   MetadataFormChoiceListValueYAML,
   MetadataPrimitiveValueType,
@@ -63,6 +64,56 @@ export const exportMetadataValueToYAML = (
 
   const handler = primitiveValueHandlers[data.type as MetadataPrimitiveValueType]
   return handler.toYAML(context, data)
+}
+
+const exportFixedArrayElementToYAML = (
+  context: ConfigurationContext,
+  value: MetadataTypedValue | undefined
+): MetadataFixedArrayValueYAML[number] => {
+  if (value === undefined) return undefined
+  if (value.type === "formChoiceListDesTimeValue") {
+    return {
+      Тип: "ЗначениеСпискаВыбора",
+      ...exportFormChoiceListToYAML(context, value),
+    }
+  }
+  return exportMetadataValueToYAML(context, undefined, value) as MetadataFixedArrayValueYAML[number]
+}
+
+export const exportFixedArrayToYAML = (
+  context: ConfigurationContext,
+  data: MetadataFixedArrayValue
+): MetadataFixedArrayValueYAML =>
+  data.value.map((value) => exportFixedArrayElementToYAML(context, value)) as MetadataFixedArrayValueYAML
+
+const exportExplicitChoiceListValueToYAML = (
+  value: MetadataFormChoiceListValue["value"]
+): MetadataFormChoiceListValueYAML["Значение"] | undefined => {
+  if (value?.type !== "DataCompositionComparisonType") return undefined
+  return {
+    Тип: "ВидСравненияКомпоновкиДанных",
+    Значение: DataCompositionComparisonTypeToYAML[
+      value.value as keyof typeof DataCompositionComparisonTypeToYAML
+    ],
+  }
+}
+
+export const exportFormChoiceListToYAML = (
+  context: ConfigurationContext,
+  data: MetadataFormChoiceListValue
+): MetadataFormChoiceListValueYAML => {
+  const valueResult =
+    exportExplicitChoiceListValueToYAML(data.value) ??
+    exportMetadataValueToYAML(context, undefined, data.value as MetadataTypedValue | undefined)
+  const presentation = exportI8nTextToYAML({
+    context,
+    rule: { type: "I8nText" },
+    value: data.presentation,
+  })
+  const result: MetadataFormChoiceListValueYAML = {}
+  if (presentation !== undefined) result.Представление = presentation
+  if (valueResult !== undefined) result.Значение = valueResult
+  return result
 }
 
 /** @deprecated Используй exportFormChoiceListToYAML из submodule formChoiceList/toYAML */
